# tokenizers/sentencepiece: Improve performance by removing allocations